Full Force!!!
Now we type out Kemmi's whole dialogue set with ease!!!

But first... create a DialogueNode. You should put in a value for what Dialogue Set to read, as well as have a list of events to do at end of the set, and toggle to run on awake or not.